requirements.txt
fastapi
uvicorn
pandas
openai
langchain
python-dotenv
plotly
matplotlib
python-dotenv


